moved keypair io methods to WalletManager
