+++ distributeblocks/Wallet.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import distributeblocks.crypto.*;
+import distributeblocks.io.WalletManager;
 
 
 /*
@@ -20,7 +21,7 @@ public class Wallet {
 	// Coin base keys are used for signing block reward transactions from a static source
 	private static final String COIN_BASE_ID = "COIN_BASE";
 	private static final String COIN_BASE_DIR = System.getProperty("user.dir") + "/coinBase";
-	private static final KeyPair COIN_BASE_KEYS = Crypto.loadKeyPair(COIN_BASE_DIR, Crypto.GEN_ALGORITHM);
+	private static final KeyPair COIN_BASE_KEYS = WalletManager.loadKeyPair(COIN_BASE_DIR, Crypto.GEN_ALGORITHM);
 	private static final float BLOCK_REWARD_AMOUNT = 5.0f;
 
 	private PrivateKey privateKey;
